pyproject.toml
[project]
name = "ib"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "reservoirinvest", email = "reservoirinvest@gmail.com"},
]
dependencies = [
    "numpy>=2.1.2",
    "pandas>=2.2.3",
    "pytz>=2024.2",
    "pyyaml>=6.0.2",
    "ib-async>=1.0.3",
    "loguru>=0.7.2",
    "scipy>=1.14.1",
    "tabulate>=0.9.0",
    "jupyterlab>=4.2.5",
    "tqdm>=4.66.5",
    "ipykernel>=6.29.5",
    "ruff>=0.6.9",
    "click>=8.1.7",
    "from-root>=1.3.0",
    "python-dotenv>=1.0.1",
    "lxml>=5.3.0",
    "yfinance>=0.2.44",
    "ibapi>=9.81.1.post1",
    "vulture>=2.13",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
